ats_ai_scanner/requirements.txt
groq
python-dotenv
pdfminer.six
spacy
sentence-transformers
scikit-learn
